+++ photutils/segmentation/catalog.py
@@ -357,6 +357,7 @@ class SourceCatalog:
 
         self.default_columns = DEFAULT_COLUMNS
         self._extra_properties = []
+        self._fluxfrac_cache = {}
         self.meta = _get_meta()
         self._update_meta()
 
@@ -525,6 +526,11 @@ class SourceCatalog:
         if not newcls.isscalar:
             value = value.tolist()
         setattr(newcls, attr, value)
+
+        # Slice the fluxfrac_radius cache values
+        newcls._fluxfrac_cache = {key: value[index]
+                                  for key, value
+                                  in self._fluxfrac_cache.items()}
 
         # evaluated lazyproperty objects and extra properties
         keys = (set(self.__dict__.keys())
@@ -1379,7 +1385,7 @@ class SourceCatalog:
         within a circular aperture of ``4 * sigma`` from the current
         position, weighting pixel values with a 2D Gaussian with a
         standard deviation of ``sigma``. ``sigma`` is the half-light
-        radius (i.e., ``flucfrac_radius(0.5)``) times (2.0 / 2.35). A
+        radius (i.e., ``fluxfrac_radius(0.5)``) times (2.0 / 2.35). A
         minimum half-light radius of 0.5 pixels is used. Iteration stops
         when the change in centroid position falls below a pre-defined
         threshold or a maximum number of iterations is reached.
@@ -3624,6 +3630,13 @@ class SourceCatalog:
             msg = 'fluxfrac must be > 0 and <= 1'
             raise ValueError(msg)
 
+        # return cached result if available
+        if fluxfrac in self._fluxfrac_cache:
+            result = self._fluxfrac_cache[fluxfrac]
+            if name is not None:
+                self.add_extra_property(name, result, overwrite=overwrite)
+            return result
+
         args = self._fluxfrac_optimizer_args
         if self.progress_bar:  # pragma: no cover
             desc = 'fluxfrac_radius'
@@ -3674,6 +3687,7 @@ class SourceCatalog:
             radius.append(result)
 
         result = np.array(radius) << u.pix
+        self._fluxfrac_cache[fluxfrac] = result
 
         if name is not None:
             self.add_extra_property(name, result, overwrite=overwrite)